requirements.txt
flask
mysql-connector-python
python-dotenv
itsdangerous==2.0.1
